requirements.txt
python-telegram-bot==13.15
tweepy>=4.14.0
openai>=1.3.5

